Refactor utils
Create EquiTrack utils dir, move current utils files there
Move utils/common/utils.py functions to EquiTrack
