+++ EquiTrack/EquiTrack/settings/local.py
@@ -98,5 +98,5 @@ LOGGING['handlers']['console']['filters'] = ['tenant_context']
 LOGGING['handlers']['console']['formatter'] = 'tenant_context'
 
 SHELL_PLUS_PRE_IMPORTS = (
-    ('EquiTrack.util_scripts', '*'),
+    ('EquiTrack.utils.scripts', '*'),
 )
