+++ EquiTrack/management/issues/checks.py
@@ -10,7 +10,7 @@ from django.utils.module_loading import import_string
 from environment.models import IssueCheckConfig
 from management.issues.exceptions import IssueCheckNotFoundException, IssueFoundException
 from management.models import FlaggedIssue, ISSUE_STATUS_RESOLVED
-from utils.common.utils import run_on_all_tenants
+from EquiTrack.utils import run_on_all_tenants
 
 ModelCheckData = namedtuple('ModelCheckData', 'object metadata')
